[#14] feat: 셀 디자인 변경
