+++ iOS/Baseball/Baseball/Views/MatchCell.swift
@@ -8,8 +8,31 @@
 
 import UIKit
 
+@IBDesignable
 class MatchCell: UICollectionViewCell {
     
+    @IBInspectable var cornerRadius: CGFloat = 0 {
+        didSet { configureView() }
+    }
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        configureView()
+    }
+    
+    required init?(coder: NSCoder) {
+        super.init(coder: coder)
+        configureView()
+    }
+    
+    private func configureView() {
+        contentView.layer.cornerRadius = cornerRadius
+        layer.shadowColor = UIColor.black.cgColor
+        layer.shadowOffset = CGSize(width: 5, height: 5)
+        layer.shadowRadius = 2.0
+        layer.shadowOpacity = 0.2
+        layer.masksToBounds = false
+    }
 }
 
 extension MatchCell: ReusableView { }
